inago-stream/requirements.txt
python-socketio
websocket-server
websocket-client
pika==1.1.0
zmq
